Major Changes
Removed the initial login.  Implemented Anonymous login, and allow a user to signup / login at a later time.
If a user decides to login, then it deletes the previous anon user that was created.  Todo - maybe check if data needs to be merged between the users.
Updated the navigation drawer to reflect this change.
Also moved "Premium Upgrade" to the top "secondary" option in the drawer.

Removed the Parse Login UI module
Removed the Facebook API

Updated to Parse 1.9.0

Workout Log now updates when return from the settings activity